% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_beta_quasi.R, R/hello.R
\name{gee_beta_quasi}
\alias{gee_beta_quasi}
\title{Quasi-likelihood GEE for Beta regression}
\usage{
gee_beta_quasi(
  formula,
  id,
  data,
  link = "logit",
  corstr = "independence",
  alpha = 0,
  maxit = 50,
  tol = 1e-06
)

gee_beta_quasi(
  formula,
  id,
  data,
  link = "logit",
  corstr = "independence",
  alpha = 0,
  maxit = 50,
  tol = 1e-06
)
}
\arguments{
\item{formula}{model formula}

\item{id}{cluster id}

\item{data}{data frame}

\item{link}{link function (default logit)}

\item{corstr}{working correlation structure}

\item{alpha}{correlation parameter}

\item{maxit}{maximum iterations}

\item{tol}{convergence tolerance}

\item{y}{Response in (0,1)}

\item{X}{Design matrix}
}
\value{
A list with estimated coefficients and dispersion

An object of class \code{geeBetaQuasi}
}
\description{
Fits a beta regression model for longitudinal data
using quasi-likelihood and generalized estimating equations.
}
\references{
Liang, K.-Y. and Zeger, S. L. (1986).
Longitudinal data analysis using generalized linear models.
Biometrika.
}
