% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{R_work}
\alias{R_work}
\title{Working correlation matrix}
\usage{
R_work(Ti, alpha = 0, type = "independence")
}
\arguments{
\item{Ti}{Cluster size}

\item{alpha}{Correlation parameter}

\item{type}{Correlation structure}
}
\description{
Working correlation matrix
}
